refactor tests and docs
